Added pak0.pak and source files from other repos.
In the old subversion repositories, there were two instances of cq3d (cq3d and
cq3dnew). The only one that _worked_ was cq3dnew, but that didn't have any
history, so this is the one revision from cq3dnew applied to the last revision
of the old cq3d svn repos.